# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Productos",
        "description": "Documentación de la API RESTful para gestión de productos y precios en distintas divisas",
        "contact": {
            "name": "José Miguel Dueñas Torres",
            "email": "jomiguel097@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/currencies": {
            "get": {
                "tags": [
                    "Divisas"
                ],
                "summary": "Obtener lista de divisas",
                "operationId": "73e0a13248f9f6aea618716227eab68a",
                "responses": {
                    "200": {
                        "description": "Lista de divisas obtenida exitosamente"
                    }
                }
            },
            "post": {
                "tags": [
                    "Divisas"
                ],
                "summary": "Crear nueva divisa",
                "operationId": "57e1c3f26ae9c82bd5996ac5bb738c5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "symbol",
                                    "exchange_rate"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dólar"
                                    },
                                    "symbol": {
                                        "type": "string",
                                        "example": "$"
                                    },
                                    "exchange_rate": {
                                        "type": "number",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Divisa creada exitosamente"
                    }
                }
            }
        },
        "/api/currencies/{id}": {
            "get": {
                "tags": [
                    "Divisas"
                ],
                "summary": "Obtener una divisa por ID",
                "operationId": "23440d4caad0c601f680b655f9c907f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Divisa encontrada"
                    },
                    "404": {
                        "description": "Divisa no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Divisas"
                ],
                "summary": "Actualizar una divisa",
                "operationId": "8a397517013e49ac5bf2db204caaa130",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Peso"
                                    },
                                    "symbol": {
                                        "type": "string",
                                        "example": "$"
                                    },
                                    "exchange_rate": {
                                        "type": "number",
                                        "example": 4000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Divisa actualizada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Divisas"
                ],
                "summary": "Eliminar una divisa",
                "operationId": "c5d45f11f33d23039acf71db15a4cf31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Divisa eliminada"
                    },
                    "404": {
                        "description": "Divisa no encontrada"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener lista de productos",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida exitosamente"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "currency_id",
                                    "tax_cost",
                                    "manufacturing_cost"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Producto A"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción del producto"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "currency_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tax_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "manufacturing_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener un producto por ID",
                "operationId": "7ab21ddb713501f3d4c2914892c02a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto",
                "operationId": "6a7e11900844cd85cff66e662b82bd4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Producto actualizado"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción actualizada"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 120
                                    },
                                    "currency_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tax_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 12
                                    },
                                    "manufacturing_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 22
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado exitosamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto",
                "operationId": "e80a6ca46716acee6b47ea1ac91663c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado exitosamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            }
        },
        "/api/products/{id}/prices": {
            "get": {
                "tags": [
                    "Precios de productos"
                ],
                "summary": "Obtener precios del producto en diferentes divisas",
                "operationId": "c344001ef9e38ce3a9a475ff2f6de5a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de precios obtenida"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Precios de productos"
                ],
                "summary": "Registrar nuevo precio en divisa para producto",
                "operationId": "0a267ed14eb16d878bbec2d44c48cc7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "currency_id",
                                    "price"
                                ],
                                "properties": {
                                    "currency_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 95.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Precio registrado correctamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Divisas",
            "description": "Divisas"
        },
        {
            "name": "Productos",
            "description": "Productos"
        },
        {
            "name": "Precios de productos",
            "description": "Precios de productos"
        }
    ]
}